+++ crates/chainlink-datastreams/src/gmsol.rs
@@ -43,7 +43,7 @@ impl super::FromChainlinkReport for PriceFeedPrice {
 
         let observations_timestamp = report.observations_timestamp;
 
-        let last_update_diff_ns = if let Some(last_update_timestamp_ns) =
+        let last_update_diff_secs = if let Some(last_update_timestamp_ns) =
             report.last_update_timestamp()
         {
             let observations_timestamp_ns = u64::from(observations_timestamp)
@@ -66,10 +66,10 @@ impl super::FromChainlinkReport for PriceFeedPrice {
                         0
                     }
                 };
-            let diff = match u32::try_from(last_update_diff) {
+            let diff = match u32::try_from(last_update_diff.div_ceil(NANOS_PER_SECOND)) {
                 Ok(diff) => diff,
                 Err(_) => {
-                    // If `last_update_diff_ns` exceeds the range representable by a `u32`,
+                    // If `last_update_diff_secs` exceeds the range representable by a `u32`,
                     // we consider the data too old. According to Chainlink Data Streams'
                     // specification for `last_update_timestamp`, such a cause should be
                     // treasted as the market being closed.
@@ -88,14 +88,15 @@ impl super::FromChainlinkReport for PriceFeedPrice {
             (price / divisor).try_into().unwrap(),
             (bid / divisor).try_into().unwrap(),
             (ask / divisor).try_into().unwrap(),
-            last_update_diff_ns.unwrap_or(0),
+            last_update_diff_secs.unwrap_or(0),
         );
 
         price.set_flag(PriceFlag::Open, is_open);
         price.set_flag(
             PriceFlag::LastUpdateDiffEnabled,
-            last_update_diff_ns.is_some(),
+            last_update_diff_secs.is_some(),
         );
+        price.set_flag(PriceFlag::LastUpdateDiffSecs, true);
 
         Ok(price)
     }